refactored mapping view into separate file so it can be accessed centrally from can channels view and obd2channelsview
